requirements.txt
Django==3.1.6
django-extensions==3.1.1
django-filter==2.4.0
django-fsm==2.7.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.2
drf-base64==2.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.0.0
flake8==3.9.0
ipython==7.21.0
ipython-genutils==0.2.0
psycopg2-binary==2.8.6
pytest==6.2.2
pytest-django==4.1.0
pytest-flake8==1.0.7
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-mock==1.8.0
sentry-sdk==1.0.0
simplejson==3.17.2
swagger-spec-validator==2.7.3
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
yarl==1.6.3
